# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "plot api",
    "contact": {}
  },
  "host": "localhost:4000",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/subject/5f8c7ee1b9776e05f105a6db": {
      "get": {
        "summary": "Subject",
        "tags": ["Subject"],
        "operationId": "Subject",
        "deprecated": false,
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
...
}